draft of clean implementation of GradientMethod
