+++ src/dalia/configs/dalia_config.py
@@ -28,7 +28,6 @@ class BFGSConfig(BaseModel):
     # c1: float = 1e-4  # only relevant for BFGS not for L-BFGS-B
     # c2: float = 0.9  # only relevant for BFGS not for L-BFGS-B
     disp: bool = False
-    
 
 
 class DaliaConfig(BaseModel):
