+++ src/dalia/gradient_methods/__init__.py
@@ -1,0 +1,6 @@
+# Copyright 2024-2025 DALIA authors. All rights reserved.
+
+from dalia.gradient_methods.vanilla_gradient import VanillaGradient
+# from dalia.gradient_methods.smart_gradient import SmartGradient
+
+__all__ = ["VanillaGradient"]
